Move formatDefaultSeoData to seo subfolder
