+++ functions/wordpress/postTypes/getPostTypeTaxonomyArchive.js
@@ -1,5 +1,5 @@
-import formatDefaultSeoData from '@/functions/formatDefaultSeoData'
 import getMenus from '@/functions/wordpress/menus/getMenus'
+import formatDefaultSeoData from '@/functions/wordpress/seo/formatDefaultSeoData'
 import queryPostsByCategory from '@/lib/wordpress/categories/queryPostsByCategory'
 import {initializeWpApollo} from '@/lib/wordpress/connector'
 import queryPostsByTag from '@/lib/wordpress/tags/queryPostsByTag'
